add dto, check API
