+++ src/main/java/com/threejavers/drinkapp/dto/DrinkDTO.java
@@ -1,0 +1,41 @@
+package com.threejavers.drinkapp.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+import java.util.UUID;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class DrinkDTO {
+
+    private UUID id;
+    private String name;
+    private Boolean isCustom;
+    private Boolean isApproved;
+    private String recipe;
+    private String drinkType;
+    private String glassType;
+    private String modificationDate;
+    private String imageUrl;
+    private CategoryDTO category;
+    private List<IngredientDTO> ingredientList;
+
+    @Override
+    public String toString() {
+        return "DrinkDTO{" +
+                "name='" + name + '\'' +
+                ", isCustom=" + isCustom +
+                ", isApproved=" + isApproved +
+                ", recipe='" + recipe + '\'' +
+                ", drinkType='" + drinkType + '\'' +
+                ", glassType='" + glassType + '\'' +
+                ", modificationDate='" + modificationDate + '\'' +
+                ", imageUrl='" + imageUrl + '\'' +
+                ", category=" + category +
+                '}';
+    }
+}
