+++ src/main/java/com/threejavers/drinkapp/dto/IngredientDTO.java
@@ -1,0 +1,17 @@
+package com.threejavers.drinkapp.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class IngredientDTO {
+
+    private UUID id;
+    private String name;
+    private String measure;
+}
