+++ src/main/java/com/threejavers/drinkapp/model/DrinkList.java
@@ -1,0 +1,16 @@
+package com.threejavers.drinkapp.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class DrinkList {
+
+    private List<Drink> drinkList;
+
+}
